requirements.txt
asgiref==3.3.4
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.5
django-webpack-loader==1.1.0
djangorestframework==3.12.4
filelock==3.0.12
platformdirs==2.0.2
pycodestyle==2.7.0
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.6.0
